adding observedOnly/expectedOnly to significance calculation

+++ scripts/limit.py
@@ -519,10 +519,11 @@ for directory in args :
                 os.system("hadd batch_collected_%s.root %s/res/*.root" % (ifile, name))
                 ifile=ifile+1
         ## and finally hadd all sub files corresponding to each crab directory
-        if os.path.exists("batch_collected.root") :
-            os.system("rm batch_collected.root")
-        os.system("hadd batch_collected.root batch_collected_*.root")
-        os.system("rm batch_collected_*.root")
+        if not options.observedOnly :
+            if os.path.exists("batch_collected.root") :
+                os.system("rm batch_collected.root")
+            os.system("hadd batch_collected.root batch_collected_*.root")
+            os.system("rm batch_collected_*.root")
         ## in case there were no batch jobs run run interactively
         ## combine datacard from all datacards in this directory
         os.system("combineCards.py -S *.txt > tmp.txt")
@@ -530,13 +531,14 @@ for directory in args :
         mass_value = directory[directory.rfind("/")+1:]
         mass_fixed = options.fixed_mass if options.fixed_mass!="" else mass_value
         os.system("text2workspace.py --default-morphing=%s -m %s -b tmp.txt -o tmp.root"% (options.shape, mass_fixed))
-        if file==0 :
+        if not options.observedOnly :
             ## calculate significance, batch_collected.root is the output file name expected by plot.cc
             os.system("combine -M ProfileLikelihood -t {toys} --significance --signalForSignificance={sig} -m {mass} -n batch_collected.root tmp.root".format(
                 toys=options.toys, sig=options.signal_strength, mass=mass_value))
-        ## calc observed significance
-        print "combine -M ProfileLikelihood --significance -m {mass} tmp.root".format(mass=mass_value)
-        os.system("combine -M ProfileLikelihood --significance -m {mass} tmp.root".format(mass=mass_value))
+        if not options.expectedOnly :
+            ## calc observed significance
+            #print "combine -M ProfileLikelihood --significance -m {mass} tmp.root".format(mass=mass_value)
+            os.system("combine -M ProfileLikelihood --significance -m {mass} tmp.root".format(mass=mass_value))
     if options.prepMSSMxsec :
         ## prepare mass argument for limit calculation if configured such
         idx = directory.rfind("/")
